refactor(map): modularização do mapa separando o do world
